requirements.txt
django
djangorestframework
python-dotenv
psycopg2
Pillow